Remove MXNet dependencies from core
