+++ core/src/autogluon/core/__init__.py
@@ -5,5 +5,4 @@ from .decorator import *
 from .utils.files import *
 from .scheduler.resource.resource import *
 from .scheduler.scheduler import *
-from . import optimizer
