+++ mxnet/src/autogluon/mxnet/utils/__init__.py
@@ -1,4 +1,4 @@
 from ..utils.dataloader import *
 from ..utils.dataset import *
 from ..utils.openml_download import *
-
+from ..utils.mxutils import *
