--- mxnet/src/autogluon/mxnet/utils/try_import.py
@@ -1,22 +1,0 @@
-__all__ = ['try_import_mxnet']
-
-
-def try_import_mxnet():
-    mx_version = '1.6.0'
-    try:
-        import mxnet as mx
-        from distutils.version import LooseVersion
-
-        if LooseVersion(mx.__version__) < LooseVersion(mx_version):
-            msg = (
-                "Legacy mxnet=={} detected, some new modules will not work properly. "
-                "mxnet>={} is required. You can use pip to upgrade mxnet "
-                "`pip install mxnet --upgrade` "
-                "or `pip install mxnet_cu101 --upgrade`").format(mx.__version__, mx_version)
-            raise ValueError(msg)
-    except ImportError:
-        raise ImportError(
-            "Unable to import dependency mxnet. "
-            "A quick tip is to install via `pip install mxnet --upgrade`, "
-            "or `pip install mxnet_cu101 --upgrade`")
-
