+++ tabular/src/autogluon/tabular/models/fastainn/tabular_nn_fastai.py
@@ -16,7 +16,7 @@ from ..abstract.abstract_model import AbstractModel
 from ...constants import REGRESSION, BINARY, MULTICLASS
 from autogluon.core.utils.loaders import load_pkl
 from autogluon.core.utils.savers import save_pkl
-from ...try_import import try_import_fastai_v1
+from autogluon.core.utils import try_import_fastai_v1
 
 # FIXME: Has a leak somewhere, training additional models in a single python script will slow down training for each additional model. Gets very slow after 20+ models (10x+ slowdown)
 #  Slowdown does not appear to impact Mac OS
