+++ tabular/src/autogluon/tabular/models/lgb/hyperparameters/lgb_trial.py
@@ -3,7 +3,7 @@ import logging
 from ...abstract import model_trial
 from autogluon.core.utils.loaders import load_pkl
 from autogluon.core.utils.exceptions import TimeLimitExceeded
-from ....try_import import try_import_lightgbm
+from autogluon.core.utils import try_import_lightgbm
 from autogluon.core import args
 
 logger = logging.getLogger(__name__)
