+++ tabular/src/autogluon/tabular/models/tabular_nn/tabular_nn_model.py
@@ -32,7 +32,7 @@ from ...constants import BINARY, MULTICLASS, REGRESSION, SOFTCLASS
 from ...features.feature_metadata import R_INT, R_FLOAT, R_CATEGORY, R_OBJECT
 from ...metrics import log_loss, roc_auc
 from autogluon.core import Space
-from autogluon.core.utils.try_import import try_import_mxboard
+from autogluon.core.utils import try_import_mxboard
 
 warnings.filterwarnings("ignore", module='sklearn.preprocessing')  # sklearn processing n_quantiles warning
 logger = logging.getLogger(__name__)
