+++ vision/src/autogluon/vision/object_detection/pipeline.py
@@ -23,7 +23,7 @@ from .utils import get_lr_scheduler, get_yolo3_metrics, get_faster_rcnn_metrics,
 # TODO: move it to general util.py
 from autogluon.mxnet.task.utils import _train_val_split
 from autogluon.core import args
-from autogluon.core.utils.mxutils import collect_params
+from autogluon.mxnet.utils import collect_params
 
 
 def get_dataloader(net, train_dataset, val_dataset, data_shape, batch_size, num_workers, args):
