+++ vision/src/autogluon/vision/utils/__init__.py
@@ -1,4 +1,3 @@
-from .try_import import *
 from .learning_rate import *
 from .plot_network import plot_network
